feature: 상태 관리 구현
ㄴ
