+++ src/components/LoginPage.jsx
@@ -1,4 +1,12 @@
-const LoginPage = () => `
+import { isLoggedIn } from "../utils/login.js";
+import ProfilePage from "./ProfilePage.jsx";
+
+const LoginPage = () => {
+  if (isLoggedIn()) {
+    window.history.replaceState({}, "", "/profile");
+    return ProfilePage();
+  }
+  return `
   <main class="bg-gray-100 flex items-center justify-center min-h-screen">
     <div class="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
       <h1 class="text-2xl font-bold text-center text-blue-600 mb-8">항해플러스</h1>
@@ -21,5 +29,6 @@ const LoginPage = () => `
     </div>
   </main>
 `;
+};
 
 export default LoginPage;
